refactor: create dayoverview package
