+++ app/src/main/res/layout/fragment_day_overview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DayOverviewFragment">
+    tools:context=".dayoverview.DayOverviewFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/activityListRecyclerView"
