+++ app/src/main/res/navigation/nav_graph.xml
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/DayOverviewFragment"
-        android:name="dev.jefvda.smartplanner.DayOverviewFragment"
+        android:name="dev.jefvda.smartplanner.dayoverview.DayOverviewFragment"
         android:label="@string/day_overview_fragment_label"
         tools:layout="@layout/fragment_day_overview">
         <argument
